LayoutTests/fast/js/reparsing-semicolon-insertion-expected.txt

This test checks that automatic semicolon insertion for parsing and reparsing agree. In a debug build, this test will fail with an assertion failure if they do not.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS commaTest() is undefined.
PASS varCommaTest() is undefined.
PASS constCommaTest() is undefined.
PASS successfullyParsed is true

TEST COMPLETE
